tutorial/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --output-file=requirements.txt --resolver=backtracking requirements.in
#
asgiref==3.6.0
    # via django
backports-zoneinfo==0.2.1 ; python_version < "3.9"
    # via
    #   -r requirements.in
    #   django
django==4.1.7
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.14.0
    # via -r requirements.in
gindix-multitenant==3.1.1
    # via -r requirements.in
psycopg2==2.9.5
    # via -r requirements.in
pytz==2022.7.1
    # via djangorestframework
sqlparse==0.4.3
    # via django
